Update the groovy code
* Use the same base analsyer test
* remove the word task as the java one is as basic as it gets
* update where relevant the other tasks

+++ src/main/groovy/freeman/ollie/analysis/GroovyFileContentAnalysisTask.groovy
@@ -28,7 +28,7 @@ class GroovyFileContentAnalysisTask extends RecursiveTask<Map<Integer, Long>> {
         tasks.each {task ->
             Map<Integer, Long> taskResults = task.join()
             taskResults.each {length, count ->
-                results[length] = (results[length] ?: 0) + count
+                results[length] = ((results[length] ?: 0) + count) as Long
             }
         }
